requirements.txt
pymupdf>=1.23.0
pdfplumber>=0.10.0
Pillow>=10.0.0
numpy>=1.24.0
opencv-python-headless>=4.8.0
pytesseract>=0.3.10
